Enlarge type of argument nsteps to vm_step and vm_rstep

+++ vm/cli.pxd
@@ -94,9 +94,11 @@ cdef extern from "vm.h":
     bint vm_break(VMState *state, size_t code_offset)
     bint vm_cont(VMState *state, VMStateDiff **diff, bint *hit_bp)
     void vm_rcont(VMState *state, VMStateDiff **diff, bint *hit_bp)
-    bint vm_step(VMState *state, int nsteps, VMStateDiff **diff, bint *hit_bp)
+    bint vm_step(VMState *state, unsigned long nsteps, VMStateDiff **diff,
+                 bint *hit_bp)
     bint vm_run(VMState *state, VMStateDiff **diff, bint *hit_bp)
-    void vm_rstep(VMState *state, int nsteps, VMStateDiff **diff, bint *hit_bp)
+    void vm_rstep(VMState *state, unsigned long nsteps, VMStateDiff **diff,
+                  bint *hit_bp)
     void vm_break_async_from_signal(VMState *state)
     void vm_break_async_from_thread(VMState *state)
     bint vm_interrupt(VMState *state, VMInterruptType type, ...)
